refactor(tests): Move metrics unit tests to internal/metrics and add new tests
- Relocate metrics unit tests and helper functions from `cmd/watchdog` to `internal/metrics`
  to ensure they are included in Go's test coverage report and follow Go testing conventions.
- Add unit tests for `CountVehiclePositions` and `SetupSentry` functions.
